+++ R/copy_if_not_already_exist.R
@@ -1,69 +1,66 @@
 #' Copy file or directory if they do not already exist
 #' 
-#' @param from character. The file or directory to copy
-#' @param to character. The target path to copy to
-#' @param copy_type character. The type of element to copy (file or dir)
-#' @param quiet logical. Should the use be warned about pre-existing file/dir
+#' @param from character. The extensions directory to copy
+#' @param to character. The target path to copy to extensions to
+#' @param quiet logical. Should the use be warned about pre-existing dir
 #' 
-#' @return character. Path of added file or directory.
+#' @importFrom fs dir_ls dir_copy dir_exists path_rel
+#' @importFrom purrr walk
+#' 
+#' @return character. Path of added directory.
 #' 
 #' @noRd
 #' @examples
 #' tmpdir <- tempfile(pattern = "copy")
 #'
 #' copy_if_not_already_exist(
-#'   from = system.file("_extensions","ThinkR-open","thinkridentity", package = "squash"),
-#'   to = file.path(tmpdir, "thinkridentity"),
-#'   copy_type = "dir"
+#'   from = system.file("_extensions", package = "squash"),
+#'   to = file.path(tmpdir, "_extensions")
 #' )
 #'
 #' unlink(tmpdir, recursive = TRUE)
 copy_if_not_already_exist <- function(
     from,
     to,
-    copy_type = c("file", "dir"),
     quiet = FALSE
 ){
-  # verify input
-  copy_type <- match.arg(copy_type)
+
+  # list extensions in input dir
+  ext_list <- dir_ls(path = from, recurse = 1, type = "directory") |> 
+    path_rel(start = from)
   
-  # verify files/dir is already present
-  already_present <- file.exists(to)
+  # list extensions already present in target dir
+  already_present <- dir_exists(file.path(to, ext_list))
+  existing_ext <- ext_list[already_present]
+  new_ext <- ext_list[!already_present]
   
-  if (already_present) {
+  if (length(existing_ext) > 0) {
     if (!quiet){
       cli_alert_info(
         paste(
-          "{to} already present in quarto project.",
+          "{toString(existing_ext)} extension(s)",
+          "already present in quarto project {dirname(to)}.",
           "Using it for compil."
         )
       )
     }
-    return(NULL)
   }
   
-  if (!already_present){
-    if (copy_type == "dir"){
-      if (!dir.exists(dirname(to))){
-        dir.create(
-          dirname(to),
-          recursive = TRUE
+  if (length(new_ext) > 0){
+    
+    walk(
+      .x = new_ext,
+      .f = \(x){
+        dir_copy(path = file.path(from, x),
+                 new_path = file.path(to, x)
         )
-        to_copied_path <- dirname(to)
       }
-      
-      file.copy(
-        from = from,
-        to = dirname(to),
-        recursive = TRUE
-      )
-    } else {
-      file.copy(
-        from = from,
-        to = to
-      )
-      to_copied_path <- to
-    }
+    )
+    
+    to_copied_path <- file.path(to, new_ext)
+    
+  } else {
+    to_copied_path <- NULL
   }
   # return path that has been created
   return(to_copied_path)