+++ NEWS.md
@@ -1,3 +1,17 @@
+# squash 0.4.3
+
+* use {fs} to compute extensions relative path
+
+# squash 0.4.2
+
+* Use external extension directory.
+* Enable metadata input as yaml
+* Update doc to cimplify user example in ReadMe.
+
+# squash 0.4.1
+
+* Set rendering parameters as function parameters instead of yaml file.
+
 # squash 0.4.0
 
 * New feature : set `fix_img_path` to TRUE to enable img embedding with html code.